// lib/types/inspiration/cart.ts
export interface ProductVariant {
  /**
   *  The unique identifier for the variant.
   */
  id: string;
  /**
   * The SKU (stock keeping unit) associated with the product variant.
   */
  sku?: string;
  /**
   * The product variant’s name, or the product's name.
   */
  name: string;
  /**
   * The product variant’s price after all discounts are applied.
   */
  price: number;
  /**
   * The product variant’s price before discounts are applied.
   */
  listPrice: number;
  /**
   * Indicates if the variant is available for sale.
   */
  availableForSale?: boolean;
  /**
   * Whether a customer needs to provide a shipping address when placing
   * an order for the product variant.
   */
  requiresShipping?: boolean;
  /**
   * The image associated with the variant.
   */
  image?: Image;
  /**
   * The variant's weight. If a weight was not explicitly specified on the
   * variant, this will be the product's weight.
   */
  weight?: Measurement;
  /**
   * The variant's height. If a height was not explicitly specified on the
   * variant, this will be the product's height.
   */
  height?: Measurement;
  /**
   * The variant's width. If a width was not explicitly specified on the
   * variant, this will be the product's width.
   */
  width?: Measurement;
  /**
   * The variant's depth. If a depth was not explicitly specified on the
   * variant, this will be the product's depth.
   */
  depth?: Measurement;
}

export interface SelectedOption {
  /**
   * The unique identifier for the option.
   */
  id?: string;
  /**
   *  The product option’s name, such as "Color" or "Size".
   */
  name: string;
  /**
   * The product option’s value, such as "Red" or "XL".
   */
  value: string;
}

export interface LineItem {
  /**
   * The unique identifier for the line item.
   */
  id: string;
  /**
   * The unique identifier for the product variant.
   */
  variantId: string;
  /**
   * The unique identifier for the product, if the variant is not provided.
   */
  productId: string;
  /**
   * This is usually the product's name.
   */
  name: string;
  /**
   * The quantity of the product variant in the line item.
   */
  quantity: number;
  /**
   * List of discounts applied to the line item.
   */
  discounts: Discount[];
  /**
   * A human-friendly unique string automatically generated from the product’s name.
   */
  path: string;
  /**
   * The product variant.
   */
  variant: ProductVariant;
  /**
   * List of selected options, to be used when displaying the line item, such as Color: Red, Size: XL.
   */
  options?: SelectedOption[];
}